+++ quantum_debugger/qml/algorithms/__init__.py
@@ -5,6 +5,7 @@ VQE, QAOA, QGANs, and Quantum Reinforcement Learning implementations.
 """
 
 from .vqe import VQE
+from .vqd import VQD
 from .qaoa import QAOA
 from .qgan import QuantumGAN
 from .qrl import QuantumQLearning, SimpleEnvironment
@@ -13,6 +14,7 @@ from .dqn import QuantumDQN
 
 __all__ = [
     "VQE",
+    "VQD",
     "QAOA",
     "QuantumGAN",
     "QuantumQLearning",